requirements.txt
PyQt6>=6.4.0
Pillow>=9.0.0
reportlab>=3.6.0
PyMuPDF>=1.21.0
